Move background subtraction to Stage2
